floor plan details refactor
